Final update for page 6 (don't approve patch 1)
This is the final for page 6, therefore patch 1 is redundant and may mess stuff up if you approve it. I'll try to get rid of that pull request so you might not have to worry about it. So you're fine if you just approve this branch and the branch adding the picture of skrillex.